feat(scroll): Implement panic mode for emergency recovery
🚨 PANIC MODE IMPLEMENTATION 🚨

Added comprehensive panic mode system to ScrollEncoderManager for handling catastrophic failures:

## Panic Mode Features:
- **Automatic Triggers**:
  - Extended watchdog timeout (>10 seconds)
  - Repeated recovery failures (10+ consecutive)
  - Persistent I2C bus failure (>30 seconds)
  - Manual trigger via 10-second button press

- **Recovery Actions**:
  - PANIC_FULL_RESET: Complete system restart with NVS erase
  - PANIC_RESTORE_DEFAULTS: Reset all parameters to safe defaults
  - PANIC_BYPASS_ENCODER: Disable encoder, use default values
  - PANIC_DIAGNOSTIC_MODE: Run comprehensive diagnostics

- **Visual Feedback**:
  - Rapid red LED flashing on panic entry
  - Color-coded LED patterns for different panic states
  - Diagnostic output via serial console

## Technical Implementation:
- PanicMode class with state machine
- Aggressive I2C bus recovery (20 clock pulses)
- Memory and health status diagnostics
- Auto-exit when encoder becomes responsive
- Integration with existing health monitoring

## Code Changes:
- Added PanicMode class to ScrollEncoderManager.h
- Implemented panic logic in ScrollEncoderManager.cpp
- Updated update() method to check panic conditions
- Added manual trigger via long button press
- Updated documentation in SCROLL_ENCODER_IMPLEMENTATION.md

This provides a last-resort recovery mechanism when the scroll encoder becomes completely unresponsive, ensuring the system can always recover from catastrophic failures.

+++ src/effects/strip/LGPOrganicWavePatterns.cpp
@@ -439,7 +439,7 @@ void lgpProteinFoldingDynamics() {
     int proteinLength = 50 + (complexity * 200);  // 50-250 residues
     
     // Protein folding simulation
-    for(int i = 0; i < min(proteinLength, HardwareConfig::STRIP_LENGTH); i++) {
+    for(int i = 0; i < min(proteinLength, (int)HardwareConfig::STRIP_LENGTH); i++) {
         // Secondary structure formation (alpha helices, beta sheets)
         if (foldingProgress > 0.2f) {
             // Check for helix-forming regions
@@ -675,7 +675,7 @@ void lgpMyceliumNetworkGrowth() {
         
         // Add communication pulses
         brightness += signal * 55;
-        brightness = min(255, brightness);
+        brightness = min(255, (int)brightness);
         
         // Color based on network state
         uint8_t hue1 = gHue + (networkAge[i] * 40);  // Age gradient
